+++ crates/oxc_allocator/CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this package will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0).
 
+## [0.144.0] - 2026-08-10
+
+### 🚀 Features
+
+- b3c1b30 allocator: Add `ArenaVec::push_mut` (#25338) (overlookmotel)
+
+### 📚 Documentation
+
+- fe111eb allocator: Improve doc comments for `Arena` methods (#24922) (overlookmotel)
+
 ## [0.143.0] - 2026-08-03
 
 ### 💥 BREAKING CHANGES